requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
gunicorn==22.0.0
python-dateutil==2.9.0.post0
